Added Edit Task functionality
The user can now edit a submitted task as many times as they want.

Currently, this is done by creating a new <p> element on submitting a task and a new <input> element on editing a task every time, which results in some identical code between the addTask function and the editTask function. I plan to try a more elegant solution that toggles between the same two <p> and <input> element each time so that the editTask function can be called from inside the addTask function, like all the other functions so far (toggleCheckbox, submitTask, deleteTask...) rather than from inside the submitTask function.