Added Complete Task functionality
Merged the deleteTask and completeTask functions into one function. If a task is completed, it's stored in one array. If it's deleted, it's stored in another array. This is ultimately to add Undo functionality later on.

Also added a GIF on version 2 of this module. This version has nearly all the desired interactive behavior except for transitions between states, which will be added next.